search-mcp-openai/pyproject.toml
[project]
name = "search-mcp-openai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.0.17",
]
